8.0

+++ components/mysql_server/dynamic_loader.cc
@@ -1,4 +1,4 @@
-/* Copyright (c) 2016, Oracle and/or its affiliates. All rights reserved.
+/* Copyright (c) 2016, 2017, Oracle and/or its affiliates. All rights reserved.
 
 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -13,21 +13,27 @@ You should have received a copy of the GNU General Public License
 along with this program; if not, write to the Free Software
 Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02111-1307  USA */
 
+#include <c_string_less.h>
+#include <depth_first_search.h>
+#include <mysql/components/my_service.h>
+#include <mysql/components/service_implementation.h>
+#include <mysql/components/services/dynamic_loader.h>
+#include <mysql/components/services/dynamic_loader_scheme_file.h>
+#include <mysql/components/services/registry.h>
+#include <mysqld_error.h>
+#include <rwlock_scoped_lock.h>
+#include <scope_guard.h>
+#include <stddef.h>
 #include <functional>
 #include <map>
 #include <memory>
-#include <scope_guard.h>
 #include <set>
-#include <c_string_less.h>
-#include <depth_first_search.h>
-#include <mysql/components/services/registry.h>
-#include <mysql/components/services/dynamic_loader.h>
-#include <mysql/components/service_implementation.h>
-#include <mysql/components/my_service.h>
-#include <mysql/components/services/dynamic_loader_scheme_file.h>
-#include <mysqld_error.h>
-#include <rwlock_scoped_lock.h>
+
 #include "dynamic_loader.h"
+#include "my_dbug.h"
+#include "my_inttypes.h"
+#include "my_psi_config.h"
+#include "my_sys.h"
 #include "mysql_component.h"
 #include "registry.h"
 #include "server_component.h"
@@ -1362,7 +1368,7 @@ static void init_dynamic_loader_psi_keys(void)
   const char *category= "components";
   int count;
 
-  count= array_elements(all_dynamic_loader_rwlocks);
+  count= static_cast<int>(array_elements(all_dynamic_loader_rwlocks));
   mysql_rwlock_register(category, all_dynamic_loader_rwlocks, count);
 }
 #endif
